Examples
In the aasubmitupload, links variable that you are getting needs to be a
"a JSON encoded array of links".

And "Do not forget to urlencode (escape) all values you are submitting
them via GET request."

Auth process only needs to be done once, since it gives back 2 cookies
(if they are saved, we're good to go).